Pulled ConvertValue and ConvertValueHelper into RageUtil_ConvertValue.h
