+++ src/NotesLoaderSM.cpp
@@ -7,6 +7,7 @@
 #include "RageFileManager.h"
 #include "RageLog.h"
 #include "RageUtil.h"
+#include "RageUtil_ConvertValue.h"
 #include "Song.h"
 #include "SongManager.h"
 #include "Steps.h"
