+++ src/arch/MovieTexture/MovieTexture_FFMpeg.cpp
@@ -4,6 +4,7 @@
 #include "RageDisplay.h"
 #include "RageLog.h"
 #include "RageUtil.h"
+#include "RageUtil_ConvertValue.h"
 #include "RageFile.h"
 #include "RageSurface.h"
 #include "RageUtil_Endian.h"
